+++ src/main/kotlin/com/dobby/backend/application/usecase/SignupUseCase/email/EmailCodeSendUseCase.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.application.usecase.SignupUseCase
+package com.dobby.backend.application.usecase.SignupUseCase.email
 
 import com.dobby.backend.application.mapper.VerificationMapper
 import com.dobby.backend.application.usecase.UseCase
