--- src/main/kotlin/com/dobby/backend/presentation/api/dto/response/auth/GoogleInfoResponse.kt
@@ -1,11 +1,0 @@
-package com.dobby.backend.presentation.api.dto.response.auth
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class GoogleInfoResponse(
-    @JsonProperty("email")
-    val email: String,
-
-    @JsonProperty("name")
-    val name : String
-)
