--- src/main/kotlin/com/dobby/backend/presentation/api/dto/response/auth/GoogleTokenResponse.kt
@@ -1,8 +1,0 @@
-package com.dobby.backend.presentation.api.dto.response.auth
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class GoogleTokenResponse (
-    @JsonProperty("access_token")
-    val accessToken: String
-)
